Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7932 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/rain.cpp
@@ -47,7 +47,7 @@ Rain::Rain(irr::scene::ICameraSceneNode* camera, irr::scene::ISceneNode* parent)
     assert(m != NULL);
     
     RandomGenerator g;
-    m_next_lightning = g.get(35);
+    m_next_lightning = (float)g.get(35);
     
     for (int r=0; r<RAIN_RING_COUNT; r++)
     {
@@ -153,7 +153,7 @@ void Rain::update(float dt)
         }
         
         RandomGenerator g;
-        m_next_lightning = g.get(35);
+        m_next_lightning = (float)g.get(35);
     }
     
 }   // update
